Added basic Main() and functions to gather info
Added main to give basic formula for our SSI calculation ==> TO BE CHANGED/ UPDATED
Added notice that states our answers are somewhat hard coded, our formula needs to be adjusted to be more reliable. As of now serves as a basic method to give an idea how bad a storm will be.

Added gatheringInfo.py to give functions to gather necessary information such as Wave Height, Atmospheric Pressure, and Wind Speed.
Note that the gatheringInfo,py needs to be adjusted. Namely focus on flipping direction so latest information is grabbed not from earlier in the day. Fix WaveHeight function to gather from all stations that are in station list.
Add functionality to choose which area a storm is either hitting or will hit, either through another function or in Main function.