Adding Edited NDBC_API for functionality
Due to our project needing to grab the most recent information from the buoys, we sort by the most recent hour and minutes which the NDBC_API library does not natively support. We edit the API so that we can sort by hours & minutes, and are uploading the api to our project so someone could seamlessly install our project and have it work for them.
All credit to the creators of the NDBC_API, I take no credit for any of the code that makes this api work.